+++ test/apps/main.js
@@ -1,8 +1,8 @@
 'use strict';
 
-const test = require('jtf')
-const axios = require('axios');
+const test = require('jmr')
 
+const { axios } = test;
 
 test('get /', async t => {
    const { data } = await axios.get("/")